Update bonus files

+++ src/server_bonus.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   server.c                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: acoto-gu <acoto-gu@student.42madrid.com    +#+  +:+       +#+        */
+/*   By: acoto-gu <acoto-gu@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/15 10:00:03 by acoto-gu          #+#    #+#             */
-/*   Updated: 2023/12/05 11:13:34 by acoto-gu         ###   ########.fr       */
+/*   Updated: 2023/12/06 13:35:20 by acoto-gu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -55,9 +55,9 @@ static void	handler(int signum, siginfo_t *siginfo, void *ucontext)
 
 	(void) ucontext;
 	rx = get_rx_state();
-	if (!rx->pid_client)
+	if (!rx->pid_client && siginfo->si_pid)
 		rx->pid_client = siginfo->si_pid;
-	if (rx->pid_client != siginfo->si_pid)
+	if (siginfo->si_pid && rx->pid_client != siginfo->si_pid)
 	{
 		kill(siginfo->si_pid, SIGUSR2);
 		return ;
@@ -80,16 +80,18 @@ static void	handler(int signum, siginfo_t *siginfo, void *ucontext)
 int	main(void)
 {
 	struct sigaction	act;
+	sigset_t			block_mask;
 
-	ft_printf("Server PID: %d\n", getpid());
-	sigemptyset(&(act.sa_mask));
-	act.sa_sigaction = handler;
+	sigemptyset(&block_mask);
+	sigaddset(&block_mask, SIGINT);
+	sigaddset(&block_mask, SIGQUIT);
 	act.sa_flags = SA_RESTART | SA_SIGINFO;
+	act.sa_mask = block_mask;
+	act.sa_sigaction = handler;
 	sigaction(SIGUSR1, &act, NULL);
 	sigaction(SIGUSR2, &act, NULL);
+	ft_printf("Server PID: %d\n", getpid());
 	while (1)
-	{
 		pause();
-	}
 	return (EXIT_SUCCESS);
 }
